Add scripts for rewriting problem readmes to improve timing info
Previously, when reviewing problems, I only recorded the time it took me to implement the solution.
This does not provide an accurate picture of the total time required to solve the problem.
From now on, I will also record the time it takes me to think up the solution.
This will allow me to better track my performance on each problem and also prepare me better for real interviews/competitions.